+++ backend/windmill-api/src/variables.rs
@@ -20,7 +20,8 @@ use axum::{
 };
 use hyper::StatusCode;
 use serde_json::Value;
-use windmill_audit::{audit_log, ActionKind};
+use windmill_audit::audit_ee::audit_log;
+use windmill_audit::ActionKind;
 use windmill_common::{
     db::UserDB,
     error::{Error, JsonResult, Result},
@@ -34,6 +35,7 @@ use lazy_static::lazy_static;
 use magic_crypt::{MagicCrypt256, MagicCryptTrait};
 use serde::Deserialize;
 use sqlx::{Postgres, Transaction};
+use windmill_git_sync::{handle_deployment_metadata, DeployedObject};
 
 lazy_static! {
     pub static ref SECRET_SALT: Option<String> = std::env::var("SECRET_SALT").ok();
@@ -164,10 +166,7 @@ async fn get_variable(
                 let mc = build_crypt(&mut tx, &w_id).await?;
                 tx.commit().await?;
 
-                Some(
-                    mc.decrypt_base64_to_string(value)
-                        .map_err(|e| Error::InternalErr(e.to_string()))?,
-                )
+                Some(decrypt(&mc, value)?)
             } else if q.include_encrypted.unwrap_or(false) {
                 Some(value)
             } else {
@@ -280,6 +279,7 @@ async fn create_variable(
     Extension(db): Extension<DB>,
     Extension(user_db): Extension<UserDB>,
     Extension(webhook): Extension<WebhookShared>,
+    Extension(rsmq): Extension<Option<rsmq_async::MultiplexedRsmq>>,
     Path(w_id): Path<String>,
     Query(AlreadyEncrypted { already_encrypted }): Query<AlreadyEncrypted>,
     Json(variable): Json<CreateVariable>,
@@ -324,6 +324,18 @@ async fn create_variable(
 
     tx.commit().await?;
 
+    handle_deployment_metadata(
+        &authed.email,
+        &authed.username,
+        &db,
+        &w_id,
+        DeployedObject::Variable { path: variable.path.clone(), parent_path: None },
+        Some(format!("Variable '{}' created", variable.path.clone())),
+        rsmq.clone(),
+        true,
+    )
+    .await?;
+
     webhook.send_message(
         w_id.clone(),
         WebhookMessage::CreateVariable { workspace: w_id, path: variable.path.clone() },
@@ -352,8 +364,10 @@ async fn encrypt_value(
 
 async fn delete_variable(
     authed: ApiAuthed,
+    Extension(db): Extension<DB>,
     Extension(user_db): Extension<UserDB>,
     Extension(webhook): Extension<WebhookShared>,
+    Extension(rsmq): Extension<Option<rsmq_async::MultiplexedRsmq>>,
     Path((w_id, path)): Path<(String, StripPath)>,
 ) -> Result<String> {
     let path = path.to_path();
@@ -386,6 +400,18 @@ async fn delete_variable(
 
     tx.commit().await?;
 
+    handle_deployment_metadata(
+        &authed.email,
+        &authed.username,
+        &db,
+        &w_id,
+        DeployedObject::Variable { path: path.to_string(), parent_path: Some(path.to_string()) },
+        Some(format!("Variable '{}' deleted", path)),
+        rsmq.clone(),
+        true,
+    )
+    .await?;
+
     webhook.send_message(
         w_id.clone(),
         WebhookMessage::DeleteVariable { workspace: w_id, path: path.to_owned() },
@@ -409,9 +435,10 @@ struct AlreadyEncrypted {
 
 async fn update_variable(
     authed: ApiAuthed,
+    Extension(db): Extension<DB>,
     Extension(user_db): Extension<UserDB>,
     Extension(webhook): Extension<WebhookShared>,
-    Extension(db): Extension<DB>,
+    Extension(rsmq): Extension<Option<rsmq_async::MultiplexedRsmq>>,
     Path((w_id, path)): Path<(String, StripPath)>,
     Query(AlreadyEncrypted { already_encrypted }): Query<AlreadyEncrypted>,
     Json(ns): Json<EditVariable>,
@@ -528,6 +555,18 @@ async fn update_variable(
     .await?;
     tx.commit().await?;
 
+    handle_deployment_metadata(
+        &authed.email,
+        &authed.username,
+        &db,
+        &w_id,
+        DeployedObject::Variable { path: npath.clone(), parent_path: Some(path.to_string()) },
+        None,
+        rsmq.clone(),
+        true,
+    )
+    .await?;
+
     webhook.send_message(
         w_id.clone(),
         WebhookMessage::UpdateVariable {
@@ -595,9 +634,7 @@ pub async fn get_value_internal<'c>(
         } else if !value.is_empty() {
             let mc = build_crypt(&mut tx, &w_id).await?;
             tx.commit().await?;
-
-            mc.decrypt_base64_to_string(value)
-                .map_err(|e| Error::InternalErr(e.to_string()))?
+            decrypt(&mc, value)?
         } else {
             "".to_string()
         }
@@ -610,4 +647,9 @@ pub async fn get_value_internal<'c>(
 
 pub fn encrypt(mc: &MagicCrypt256, value: &str) -> String {
     mc.encrypt_str_to_base64(value)
+}
+
+pub fn decrypt(mc: &MagicCrypt256, value: String) -> Result<String> {
+    mc.decrypt_base64_to_string(value)
+        .map_err(|e| Error::InternalErr(e.to_string()))
 }