+++ backend/windmill-common/src/workspaces.rs
@@ -14,6 +14,11 @@ pub enum ObjectType {
     Flow,
     App,
     Folder,
+    Resource,
+    Variable,
+    Secret,
+    Schedule,
+    ResourceType,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -21,4 +26,5 @@ pub struct GitRepositorySettings {
     pub script_path: String,
     pub git_repo_resource_path: String,
     pub use_individual_branch: Option<bool>,
+    pub exclude_types_override: Option<Vec<ObjectType>>,
 }