+++ frontend/src/lib/components/apps/components/display/table/AppTable.svelte
@@ -83,6 +83,7 @@
 	$: setSearch(searchValue)
 
 	function setSearch(srch: string) {
+		$table.setPageIndex(0)
 		outputs?.search?.set(srch)
 	}
 
@@ -102,7 +103,6 @@
 	let selectedRowIndex = -1
 
 	function toggleRow(row: Record<string, any>, force: boolean = false) {
-		console.log(row)
 		let data = { ...row.original }
 		let index = data['__index']
 		delete data['__index']
@@ -140,9 +140,10 @@
 		if (searchValue === '') {
 			return result
 		}
-		return result.filter((row) =>
-			Object.values(row).some((value) => value?.toString()?.includes(searchValue))
-		)
+		return result.filter((row) => {
+			// console.log(Object.values(row).map((value) => value?.toString()))
+			return Object.values(row).some((value) => value?.toString()?.includes(searchValue))
+		})
 	}
 
 	function renderCell(x: any, props: any): any {
@@ -156,6 +157,7 @@
 	let filteredResult: Array<Record<string, any>> = []
 
 	function setFilteredResult() {
+		console.log('setting filtered result')
 		const wIndex = Array.isArray(result)
 			? (result as any[]).map((x, i) => ({ ...x, __index: i }))
 			: [{ error: 'input was not an array' }]