+++ frontend/src/lib/components/apps/editor/AppEditorHeader.svelte
@@ -1281,7 +1281,7 @@
 			>
 				<div class="flex flex-row gap-1 items-center">
 					<Bug size={14} />
-					<div> Debug runs</div>
+					<div>Debug runs</div>
 					<div class="text-2xs text-tertiary"
 						>({$jobs?.length > 99 ? '99+' : $jobs?.length ?? 0})</div
 					>